inline SyncRequester implementation
